[xmlns] test optional XML namespace on element
Currently no way to specify that the XML namespace is _required_ - which in
the case of XML Schema can be specified on a per-node basis; also it's
quite likely that a user will want to override it on a global basis for a
particular run.  For now, we will go with only one mode of operation - lax
namespaces, where the XML namespace, if specified, must match what is in
the schema, but is otherwise optional (unless there is ambiguity).