Fix bug with straight "Quant" on classes
The 'accept' function returns an empty list, if the Graph did not consume
the node it was asked to.  This case was being tested for but nothing was
being done about it, allowing garbage input to pass through (and be
discarded).  Make this case an explicit error.

Lo and behold, this reveals that several tests in t/11-elements.t had not
been updated since the Octothorpe test modules changed.  Also fix these
now failing tests.