[xmlns] test the "more namespaces than types" case
Support XMLNS wildcarding on elements, where an extra attribute is
required to store the namespaces.  Add the xmlns_attr attribute to
elements, and deal with it appropriately.

Shake up the mess that is PRANG::Graph::Meta::Element::build_graph_node()
a little bit, it's still a bit hairy but all tests pass!