Add test for 'Whatever' class and associated bugfixes
The 'accept' function on the Graph nodes must always consistently return
the same shape of values.  So, when building the graph, they must all
either have the 'name_attr' and 'xmlns_attr' values set so that they do
this.  This also requires 'Text' to remember and return these values,
though for Text nodes they are always the empty string.  Also tidy a
couple of corner cases in the 'Element' node_ok and output functions.