Whatever: allow non-Whatever children
An important use case, this allows for any old PRANG object to be thrown
into a Whatever field, and have it serialize as if it were appearing in the
place where it was normally expected.

Of course this means that they won't round-trip from YAML to XML and back
exactly; allow individual tests to skip this.

As indicated in the source, this is making a type error legal; if we are
fed object structures which violate their type constraints (as the
05-whatever-object.yaml does, its "Ampersand" object not matching
"PRANG::Graph::Whatever|Str"), then we will happily emit documents which do
not conform to the schema.  This should be fixed at some point, with this
schema-mutating behaviour only available if explicitly enabled.