xml_min => 0 should work like xml_required => 0
Previously, xml_min was only honoured if other options were specified such
that it was used.  Instead, we should always honour it; effectively letting
the user use xml_required or xml_min to specify that an attribute is
optional.  Also, detect for xml_min and xml_max > 1 when the isa does not
specify an array, and cause an error in that case.