Ignore xsi:schemaLocation; it has no useful meaning.
xsi:schemaLocation describes "hints as to the physical location of schema
documents which may be used for ‘assessment’" - ie, it is a hint to a local
validator as to which schema file contains the XML Schema definition.  On
XML documents which are transmitted, it is somewhere between a potential
security risk and totally worthless.  See
http://www.w3.org/TR/xmlschema-1/#schema-loc for the full story.