Element trait: warn if 'required' is not specified where it should be
If we must always have one value in the slot, then it should be marked as
'required' - raise a warning if that is not the case and let the user fix
it.  Set the attribute to 'required' as well, to catch the error earlier,
but by that time it is far too late (and use of ->make_immutable will
thwart this attempt at run-time meta-attribute modification).