Meta Element: suppress warning in certain common situations
If they have specified a default, then it doesn't matter that it's not
required.  Well most of the time it won't anyway.  Also, for fields where
simple values are allowed, undef is equivalent to the empty string so this
is possibly OK, or at least not always worth warning about.  So keep it
relatively quiet for the time being.