wr72880 - aggressively check all values using type constraints on input
Waiting for Moose to throw the exception when constructing the code led to
ugly downstream code: it requires parsing the string error or heavy
exception classes to discover the fault, and it means that the exception
message has an incomplete XPath.  We can do better than this.  We almost
always have the relevant Moose::Meta::Attribute, and hence its
Moose::Meta::TypeConstraint available.  So, we can check it as soon as the
value is encountered, and raise a much more specific and less cryptic error
message.

This doesn't interact with coercion very well; but usually coercion is used
for making life easier from Perl land, not changing the document on input.