wr77086: increased minimum Moose version, and fixed resulting issues
Newer versions of Moose have changed the API to Moose::Util::MetaRole, and
emit a warning if you use the old one. This broke the tests, and would
probably result in (very verbose) warnings being spewed out to anyone using
the command line tools with newer versions of Moose. Rather than try to
remove the warnings, the dependant Moose version was increased.