fix for moose 2.04
the behavior of ->parents on union types changed (since it was actually
broken - the ->parents method should return a list of other type
constraints which must *all* be satisfied in order for the constraint to
pass). the ->type_constraints method is what actually returns the list
of types that were union-ed together.

Signed-off-by: Jesse Luehrs <doy@tozt.net>